+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.6.6
+
+- Minor: upgrade dependencies.
+
+## v0.6.5
+
+- Breaking: now `koa-compress` will not be enabled by default.
+
 ## v0.6.3
 
 - Minor: upgrade dependencies.