+++ packages/cfs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @tigojs/cfs change log
 
+## v0.4.2
+
+- Fix: returns status 404 to the request with wrong file type or wrong file name.
+
 ## v0.4.1
 
 - Breaking change: In database table, we're using UUID as primary key now, and `uid` has been changed to `scopeId`.